break out elements
