+++ components/form-builder/layout/index.ts
@@ -11,4 +11,4 @@ export { Publish } from "./Publish";
 export { Published } from "./Published";
 export { Share } from "./Share";
 export { TestDataDelivery } from "./TestDataDelivery";
-export { ElementPanel } from "../panel/ElementPanel";
+export { Elements } from "../panel";
