+++ components/form-builder/panel/PanelActionsLocked.tsx
@@ -1,9 +1,10 @@
 import React from "react";
 import styled from "styled-components";
 import { useTranslation } from "next-i18next";
-import { useTemplateStore } from "../store/useTemplateStore";
+
+import { useTemplateStore } from "../store";
 import { LockIcon } from "../icons";
-import { Button } from "../shared/Button";
+import { Button } from "../shared";
 
 const Actions = styled.div`
   position: relative;
