+++ components/form-builder/panel/index.ts
@@ -1,3 +1,4 @@
+export { Elements } from "./Elements";
 export { ElementPanel } from "./ElementPanel";
 export { PanelActions } from "./PanelActions";
 export { PanelActionsLocked } from "./PanelActionsLocked";
@@ -7,3 +8,4 @@ export { Modal, ModalButton } from "./Modal";
 export { ModalForm } from "./ModalForm";
 export { PrivacyDescription } from "./PrivacyDescription";
 export { SelectedElement, getSelectedOption } from "./SelectedElement";
+export { Form } from "./Form";
