+++ pages/form-builder/edit/[[...params]].tsx
@@ -3,14 +3,14 @@ import { useTranslation } from "next-i18next";
 import { NextPageWithLayout } from "../../_app";
 import { PageProps } from "@lib/types";
 import { getServerSideProps } from "../index";
-import { ElementPanel, EditNavigation, Template, PageTemplate } from "@formbuilder/layout";
+import { Elements, EditNavigation, Template, PageTemplate } from "@formbuilder/layout";
 
 const Page: NextPageWithLayout<PageProps> = () => {
   const { t } = useTranslation("form-builder");
   const title = `${t("gcFormsEdit")} — ${t("gcForms")}`;
   return (
     <PageTemplate title={title} navigation={<EditNavigation />}>
-      <ElementPanel />
+      <Elements />
     </PageTemplate>
   );
 };
